Changed all size types to int in serilization lib.

+++ CPU/libs/serialize.h
@@ -4,7 +4,7 @@
 #include <string.h>
 #include <stdlib.h>
 
-int serialize_data(void *object, size_t nBytes, void **buffer, size_t *lastIndex);
-int deserialize_data(void *object, size_t nBytes, void *serialized_data, size_t *lastIndex);
+int serialize_data(void *object, size_t nBytes, void **buffer, int *lastIndex);
+int deserialize_data(void *object, size_t nBytes, void *serialized_data, int *lastIndex);
 
 #endif //SERIALIZATION_SERIALIZE_H
